seismology/api/requirements.txt
flask
python-dotenv
flask-restful
flask_sqlalchemy
flask-jwt-extended
flask-mail

flask_marshmallow
marshmallow-sqlalchemy
flask-filter
sqlalchemy-filters
Flask-APScheduler
requests



